requirements.txt
azure-identity
openai
mcp
mcp[cli]
azure-search-documents